Revert "Allow for ability to auto adapt Current Vulkan Version/Maximum API level based on supported API Driver version"
This reverts commit b45bb8516a82ac26db87ace56062aab6d4e29f6e.

Revert "Cleanup casts + Temp lazy fix for Double Free Buffer Access Violation Crash"

This reverts commit d3e93b3ec47c978a8af05f01b57d17eb5d681300.

Revert "Move QueueFamilyIndices and SwapChainSupportDetails Inner Classes to static initialisation"

This reverts commit 792da957b2e58cbba0b21862be4e516a85b427f2.

Revert "Cleanup casts + Temp lazy fix for Double Free Buffer Access Violation Crash"

This reverts commit d3e93b3ec47c978a8af05f01b57d17eb5d681300.

Revert "Decode Driver Version (not Guaranteed to work)"

This reverts commit 0a7234062cb44739cb2d9597bdd855d1f8b5914f.

Revert "Extension loading/Enumeration test"

This reverts commit b02d004752db68e119a9fca371405e03699ae977.

Revert "Fix VSync Toggle bug when Entering/Leaving FullScreen + Matrix Glitches with Player Model in 1st person View/Perspective"

This reverts commit 11741f1725ff871b6fddbffc66466bd9adb1f5f9.

Revert "Attempt to Optimise BakedQuad + Try to optimise Matrix operations through FMA Intrinsics"

This reverts commit a6273964591c98b11349386c8236440669fa6884.

Revert "Attempt to Disable Debugging and Runtime Checks + Add Math Optimisations (With FMA)"

This reverts commit 9f73c577b8e750bfc814efb74d5e483a61c21e92.

Revert "Allow for ability to auto adapt Current Vulkan Version/Maximum API level based on supported API Driver version"

This reverts commit b45bb8516a82ac26db87ace56062aab6d4e29f6e.

+++ src/main/java/net/vulkanmod/mixin/SplashOverlayM.java
@@ -76,7 +76,7 @@ public abstract class SplashOverlayM {
             if (this.client.currentScreen != null && g < 1.0f) {
                 this.client.currentScreen.render(matrices, mouseX, mouseY, delta);
             }
-            k = MathHelper.ceil(MathHelper.clamp(g, 0.15, 1.0) * 255.0);
+            k = MathHelper.ceil(MathHelper.clamp((double)g, 0.15, 1.0) * 255.0);
             SplashOverlay.fill(matrices, 0, 0, i, j, withAlpha(BRAND_ARGB.getAsInt(), k));
             h = MathHelper.clamp(g, 0.0f, 1.0f);
         } else {
@@ -94,7 +94,7 @@ public abstract class SplashOverlayM {
         }
         k = (int)((double)this.client.getWindow().getScaledWidth() * 0.5);
         int m = (int)((double)this.client.getWindow().getScaledHeight() * 0.5);
-        double n = Math.min((double)this.client.getWindow().getScaledWidth() * 0.75, this.client.getWindow().getScaledHeight()) * 0.25;
+        double n = Math.min((double)this.client.getWindow().getScaledWidth() * 0.75, (double)this.client.getWindow().getScaledHeight()) * 0.25;
         int p = (int)(n * 0.5);
         double d = n * 4.0;
         int q = (int)(d * 0.5);
